+++ app/controllers/returns/ManufacturedPlasticPackagingController.scala
@@ -14,51 +14,51 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.returns
 
 import cacheables.ObligationCacheable
 import connectors.CacheConnector
 import controllers.actions._
-import forms.RecycledPlasticPackagingWeightFormProvider
-
-import javax.inject.Inject
+import controllers.routes
+import forms.ManufacturedPlasticPackagingFormProvider
 import models.Mode
 import models.returns.TaxReturnObligation
 import navigation.Navigator
-import pages.RecycledPlasticPackagingWeightPage
+import pages.ManufacturedPlasticPackagingPage
 import play.api.i18n.{I18nSupport, MessagesApi}
 import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
-import views.html.RecycledPlasticPackagingWeightView
+import views.html.ManufacturedPlasticPackagingView
 
+import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
-class RecycledPlasticPackagingWeightController @Inject()(
-                                                          override val messagesApi: MessagesApi,
-                                                          cacheConnector: CacheConnector,
-                                                          navigator: Navigator,
-                                                          identify: IdentifierAction,
-                                                          getData: DataRetrievalAction,
-                                                          requireData: DataRequiredAction,
-                                                          formProvider: RecycledPlasticPackagingWeightFormProvider,
-                                                          val controllerComponents: MessagesControllerComponents,
-                                                          view: RecycledPlasticPackagingWeightView
-                                                        )(implicit ec: ExecutionContext)
-  extends FrontendBaseController with I18nSupport {
+class ManufacturedPlasticPackagingController @Inject() (
+  override val messagesApi: MessagesApi,
+  cacheConnector: CacheConnector,
+  navigator: Navigator,
+  identify: IdentifierAction,
+  getData: DataRetrievalAction,
+  requireData: DataRequiredAction,
+  formProvider: ManufacturedPlasticPackagingFormProvider,
+  val controllerComponents: MessagesControllerComponents,
+  view: ManufacturedPlasticPackagingView
+)(implicit ec: ExecutionContext)
+    extends FrontendBaseController with I18nSupport {
 
   val form = formProvider()
 
   def onPageLoad(mode: Mode): Action[AnyContent] =
     (identify andThen getData andThen requireData).async {
       implicit request =>
-        val preparedForm = request.userAnswers.get(RecycledPlasticPackagingWeightPage) match {
-          case None => form
+        val preparedForm = request.userAnswers.get(ManufacturedPlasticPackagingPage) match {
+          case None        => form
           case Some(value) => form.fill(value)
         }
+
         request.userAnswers.get[TaxReturnObligation](ObligationCacheable) match {
           case Some(obligation) => Future.successful(Ok(view(preparedForm, mode, obligation)))
-          case None => Future.successful(Redirect(routes.IndexController.onPageLoad))
-
+          case None             => Future.successful(Redirect(controllers.routes.IndexController.onPageLoad))
         }
     }
 
@@ -66,19 +66,21 @@ class RecycledPlasticPackagingWeightController @Inject()(
     (identify andThen getData andThen requireData).async {
       implicit request =>
         val pptId: String = request.pptReference
+
         val obligation = request.userAnswers.get[TaxReturnObligation](ObligationCacheable).getOrElse(
           throw new IllegalStateException("Must have an obligation to Submit against")
         )
+
         form.bindFromRequest().fold(
           formWithErrors => Future.successful(BadRequest(view(formWithErrors, mode, obligation))),
           value =>
             for {
               updatedAnswers <- Future.fromTry(
-                request.userAnswers.set(RecycledPlasticPackagingWeightPage, value)
+                request.userAnswers.set(ManufacturedPlasticPackagingPage, value)
               )
               _ <- cacheConnector.set(pptId, updatedAnswers)
             } yield Redirect(
-              navigator.nextPage(RecycledPlasticPackagingWeightPage, mode, updatedAnswers)
+              navigator.nextPage(ManufacturedPlasticPackagingPage, mode, updatedAnswers)
             )
         )
     }
