+++ app/controllers/returns/StartYourReturnController.scala
@@ -14,15 +14,13 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.returns
 
 import cacheables.ObligationCacheable
 import connectors.CacheConnector
 import controllers.actions._
 import controllers.helpers.TaxReturnHelper
 import forms.StartYourReturnFormProvider
-
-import javax.inject.Inject
 import models.{Mode, UserAnswers}
 import navigation.Navigator
 import pages.StartYourReturnPage
@@ -31,6 +29,7 @@ import play.api.mvc.{Action, AnyContent, MessagesControllerComponents}
 import uk.gov.hmrc.play.bootstrap.frontend.controller.FrontendBaseController
 import views.html.StartYourReturnView
 
+import javax.inject.Inject
 import scala.concurrent.{ExecutionContext, Future}
 
 class StartYourReturnController @Inject()(
