+++ app/views/ConfirmPlasticPackagingTotalView.scala.html
@@ -35,6 +35,6 @@
 
     @govukSummaryList(list)
 
-    @button("site.button.continue", routes.DirectlyExportedComponentsController.onPageLoad(NormalMode).url)
+    @button("site.button.continue", controllers.returns.routes.DirectlyExportedComponentsController.onPageLoad(NormalMode).url)
 
 }
