+++ app/views/HumanMedicinesPlasticPackagingWeightView.scala.html
@@ -35,7 +35,7 @@
 
     @sectionHeader(messages("humanMedicinesPlasticPackaging.caption"))
 
-    @formHelper(action = routes.HumanMedicinesPlasticPackagingWeightController.onSubmit(mode), 'autoComplete -> "off") {
+    @formHelper(action = controllers.returns.routes.HumanMedicinesPlasticPackagingWeightController.onSubmit(mode), 'autoComplete -> "off") {
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form))
