+++ app/views/ImportedPlasticPackagingView.scala.html
@@ -46,7 +46,7 @@
 
     @sectionHeader(messages("caption.total.plastic"))
 
-    @formHelper(action = routes.ImportedPlasticPackagingController.onSubmit(mode), 'autoComplete -> "off") {
+    @formHelper(action = controllers.returns.routes.ImportedPlasticPackagingController.onSubmit(mode), 'autoComplete -> "off") {
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form))
