+++ app/views/ImportedPlasticPackagingWeightView.scala.html
@@ -50,7 +50,7 @@
 
     @sectionHeader(messages("caption.total.plastic"))
 
-    @formHelper(action = routes.ImportedPlasticPackagingWeightController.onSubmit(mode), 'autoComplete -> "off") {
+    @formHelper(action = controllers.returns.routes.ImportedPlasticPackagingWeightController.onSubmit(mode), 'autoComplete -> "off") {
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form))
