+++ app/views/IndexView.scala.html
@@ -167,7 +167,7 @@
 @returnsLink = @{
     obligations.flatMap { obligations =>
         if(obligations.displaySubmitReturnsLink) {
-            Some(link(id = "start-date-return-link", text = messages("account.homePage.card.makeReturn.line3.createLink"), call = routes.StartYourReturnController.onPageLoad(NormalMode)))
+            Some(link(id = "start-date-return-link", text = messages("account.homePage.card.makeReturn.line3.createLink"), call = controllers.returns.routes.StartYourReturnController.onPageLoad(NormalMode)))
         } else {
             None
         }
