+++ app/views/ManufacturedPlasticPackagingWeightView.scala.html
@@ -51,7 +51,7 @@
 
     @sectionHeader(messages("caption.total.plastic"))
 
-    @formHelper(action = routes.ManufacturedPlasticPackagingWeightController.onSubmit(mode), 'autoComplete -> "off") {
+    @formHelper(action = controllers.returns.routes.ManufacturedPlasticPackagingWeightController.onSubmit(mode), 'autoComplete -> "off") {
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form))
