+++ app/views/RecycledPlasticPackagingWeightView.scala.html
@@ -33,7 +33,7 @@
 
     @sectionHeader(taxReturn)
 
-    @formHelper(action = routes.RecycledPlasticPackagingWeightController.onSubmit(mode), 'autoComplete -> "off") {
+    @formHelper(action = controllers.returns.routes.RecycledPlasticPackagingWeightController.onSubmit(mode), 'autoComplete -> "off") {
 
         @if(form.errors.nonEmpty) {
             @govukErrorSummary(ErrorSummaryViewModel(form))
