+++ test/navigation/NavigatorSpec.scala
@@ -18,7 +18,9 @@ package navigation
 
 import base.SpecBase
 import cacheables.AmendSelectedPeriodKey
-import controllers.routes
+import controllers.returns.{routes => returnsRoutes}
+import controllers.amends.{routes => amendsRoutes}
+
 import pages._
 import models._
 import pages.amends.{AmendAreYouSurePage, AmendDirectExportPlasticPackagingPage, AmendHumanMedicinePlasticPackagingPage, AmendImportedPlasticPackagingPage, AmendManufacturedPlasticPackagingPage, AmendRecycledPlasticPackagingPage}
@@ -59,7 +61,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(StartYourReturnPage,
               NormalMode,
               answers.get
-            ) mustBe routes.ManufacturedPlasticPackagingController.onPageLoad(NormalMode)
+            ) mustBe returnsRoutes.ManufacturedPlasticPackagingController.onPageLoad(NormalMode)
 
           }
 
@@ -70,7 +72,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(StartYourReturnPage,
               NormalMode,
               answers.get
-            ) mustBe routes.NotStartOtherReturnsController.onPageLoad
+            ) mustBe returnsRoutes.NotStartOtherReturnsController.onPageLoad
 
           }
         }
@@ -83,7 +85,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ManufacturedPlasticPackagingPage,
               NormalMode,
               answers.get
-            ) mustBe routes.ManufacturedPlasticPackagingWeightController.onPageLoad(NormalMode)
+            ) mustBe returnsRoutes.ManufacturedPlasticPackagingWeightController.onPageLoad(NormalMode)
 
           }
 
@@ -94,7 +96,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ManufacturedPlasticPackagingPage,
               NormalMode,
               answers.get
-            ) mustBe routes.ImportedPlasticPackagingController.onPageLoad(NormalMode)
+            ) mustBe returnsRoutes.ImportedPlasticPackagingController.onPageLoad(NormalMode)
 
           }
         }
@@ -107,7 +109,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ImportedPlasticPackagingPage,
               NormalMode,
               answers.get
-            ) mustBe routes.ImportedPlasticPackagingWeightController.onPageLoad(NormalMode)
+            ) mustBe returnsRoutes.ImportedPlasticPackagingWeightController.onPageLoad(NormalMode)
 
           }
 
@@ -118,7 +120,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ImportedPlasticPackagingPage,
               NormalMode,
               answers.get
-            ) mustBe routes.ConfirmPlasticPackagingTotalController.onPageLoad
+            ) mustBe returnsRoutes.ConfirmPlasticPackagingTotalController.onPageLoad
 
           }
         }
@@ -131,7 +133,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ManufacturedPlasticPackagingWeightPage,
               NormalMode,
               answers.get
-            ) mustBe routes.ImportedPlasticPackagingController.onPageLoad(NormalMode)
+            ) mustBe returnsRoutes.ImportedPlasticPackagingController.onPageLoad(NormalMode)
 
           }
 
@@ -145,7 +147,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ExportedPlasticPackagingWeightPage,
               NormalMode,
               answers.get
-            ) mustBe routes.HumanMedicinesPlasticPackagingController.onPageLoad(NormalMode)
+            ) mustBe returnsRoutes.HumanMedicinesPlasticPackagingController.onPageLoad(NormalMode)
 
           }
 
@@ -158,7 +160,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(HumanMedicinesPlasticPackagingPage,
               NormalMode,
               answers.get
-            ) mustBe routes.HumanMedicinesPlasticPackagingWeightController.onPageLoad(NormalMode)
+            ) mustBe returnsRoutes.HumanMedicinesPlasticPackagingWeightController.onPageLoad(NormalMode)
           }
 
           "navigate to RecycledPlasticPackagingWeightPage when answer is No" in {
@@ -168,7 +170,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(HumanMedicinesPlasticPackagingPage,
               NormalMode,
               answers.get
-            ) mustBe routes.RecycledPlasticPackagingWeightController.onPageLoad(NormalMode)
+            ) mustBe returnsRoutes.RecycledPlasticPackagingWeightController.onPageLoad(NormalMode)
 
           }
 
@@ -182,7 +184,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(HumanMedicinesPlasticPackagingWeightPage,
               NormalMode,
               answers.get
-            ) mustBe routes.RecycledPlasticPackagingWeightController.onPageLoad(NormalMode)
+            ) mustBe returnsRoutes.RecycledPlasticPackagingWeightController.onPageLoad(NormalMode)
 
           }
 
@@ -197,7 +199,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(RecycledPlasticPackagingWeightPage,
               NormalMode,
               answers.get
-            ) mustBe routes.ConvertedPackagingCreditController.onPageLoad(NormalMode)
+            ) mustBe returnsRoutes.ConvertedPackagingCreditController.onPageLoad(NormalMode)
 
           }
 
@@ -211,7 +213,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ConvertedPackagingCreditPage,
               NormalMode,
               answers.get
-            ) mustBe routes.ReturnsCheckYourAnswersController.onPageLoad
+            ) mustBe returnsRoutes.ReturnsCheckYourAnswersController.onPageLoad
 
           }
 
@@ -228,7 +230,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(AmendAreYouSurePage,
               NormalMode,
               answers.get
-            ) mustBe controllers.amends.routes.AmendManufacturedPlasticPackagingController.onPageLoad(NormalMode)
+            ) mustBe amendsRoutes.AmendManufacturedPlasticPackagingController.onPageLoad(NormalMode)
 
           }
 
@@ -239,7 +241,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(AmendAreYouSurePage,
               NormalMode,
               answers.get
-            ) mustBe (controllers.amends.routes.ViewReturnSummaryController.onPageLoad("TEST"))
+            ) mustBe (amendsRoutes.ViewReturnSummaryController.onPageLoad("TEST"))
           }
 
         }
@@ -249,7 +251,7 @@ class NavigatorSpec extends SpecBase {
           navigator.nextPage(AmendManufacturedPlasticPackagingPage,
             NormalMode,
             UserAnswers("id")
-          ) mustBe controllers.amends.routes.AmendImportedPlasticPackagingController.onPageLoad(NormalMode)
+          ) mustBe amendsRoutes.AmendImportedPlasticPackagingController.onPageLoad(NormalMode)
 
         }
 
@@ -258,7 +260,7 @@ class NavigatorSpec extends SpecBase {
           navigator.nextPage(AmendImportedPlasticPackagingPage,
             NormalMode,
             UserAnswers("id")
-          ) mustBe controllers.amends.routes.AmendHumanMedicinePlasticPackagingController.onPageLoad(NormalMode)
+          ) mustBe amendsRoutes.AmendHumanMedicinePlasticPackagingController.onPageLoad(NormalMode)
 
         }
 
@@ -267,7 +269,7 @@ class NavigatorSpec extends SpecBase {
           navigator.nextPage(AmendHumanMedicinePlasticPackagingPage,
             NormalMode,
             UserAnswers("id")
-          ) mustBe controllers.amends.routes.AmendDirectExportPlasticPackagingController.onPageLoad(NormalMode)
+          ) mustBe amendsRoutes.AmendDirectExportPlasticPackagingController.onPageLoad(NormalMode)
 
         }
 
@@ -276,7 +278,7 @@ class NavigatorSpec extends SpecBase {
           navigator.nextPage(AmendDirectExportPlasticPackagingPage,
             NormalMode,
             UserAnswers("id")
-          ) mustBe controllers.amends.routes.AmendRecycledPlasticPackagingController.onPageLoad(NormalMode)
+          ) mustBe amendsRoutes.AmendRecycledPlasticPackagingController.onPageLoad(NormalMode)
 
         }
 
@@ -285,7 +287,7 @@ class NavigatorSpec extends SpecBase {
           navigator.nextPage(AmendRecycledPlasticPackagingPage,
             NormalMode,
             UserAnswers("id")
-          ) mustBe controllers.amends.routes.CheckYourAnswersController.onPageLoad
+          ) mustBe amendsRoutes.CheckYourAnswersController.onPageLoad
 
         }
       }
@@ -303,7 +305,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ManufacturedPlasticPackagingPage,
               CheckMode,
               answers.get
-            ) mustBe routes.ManufacturedPlasticPackagingWeightController.onPageLoad(CheckMode)
+            ) mustBe returnsRoutes.ManufacturedPlasticPackagingWeightController.onPageLoad(CheckMode)
 
           }
 
@@ -314,7 +316,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ManufacturedPlasticPackagingPage,
               CheckMode,
               answers.get
-            ) mustBe routes.ConfirmPlasticPackagingTotalController.onPageLoad
+            ) mustBe returnsRoutes.ConfirmPlasticPackagingTotalController.onPageLoad
 
           }
         }
@@ -327,7 +329,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ImportedPlasticPackagingPage,
               CheckMode,
               answers.get
-            ) mustBe routes.ImportedPlasticPackagingWeightController.onPageLoad(CheckMode)
+            ) mustBe returnsRoutes.ImportedPlasticPackagingWeightController.onPageLoad(CheckMode)
 
           }
 
@@ -338,7 +340,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ImportedPlasticPackagingPage,
               CheckMode,
               answers.get
-            ) mustBe routes.ConfirmPlasticPackagingTotalController.onPageLoad
+            ) mustBe returnsRoutes.ConfirmPlasticPackagingTotalController.onPageLoad
 
           }
         }
@@ -351,7 +353,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ManufacturedPlasticPackagingWeightPage,
               CheckMode,
               answers.get
-            ) mustBe routes.ConfirmPlasticPackagingTotalController.onPageLoad
+            ) mustBe returnsRoutes.ConfirmPlasticPackagingTotalController.onPageLoad
 
           }
 
@@ -365,7 +367,7 @@ class NavigatorSpec extends SpecBase {
             navigator.nextPage(ImportedPlasticPackagingWeightPage,
               CheckMode,
               answers.get
-            ) mustBe routes.ConfirmPlasticPackagingTotalController.onPageLoad
+            ) mustBe returnsRoutes.ConfirmPlasticPackagingTotalController.onPageLoad
 
           }
